# contracts/openapi/operator-rest.openapi.yaml
...
openapi: 3.1.0
info:
  title: OCU Control — Operator REST
  version: 0.1.0-draft
  summary: Operator/lifecycle ingress to the OCU Control plane (operator scope only).
  description: >-
    The privileged operator/lifecycle ingress to Control: session create and destroy,
    the per-session revoke (RevokeOne), the deployment-wide DENY-ALL (RevokeAll), the
    denylist lift (LiftDeny), and the quota override (OverrideQuota). The caller is
    host-attested from the operator transport's peer credential (NFR-SEC-43); request
    bodies carry hints only. The kill-switch and force-kill routes exist ONLY on this
    ingress — the gateway service identity reaches none of them (NFR-SEC-52, NFR-SEC-26).
    Every privileged action is audit-gated before acknowledgement (NFR-SEC-45). This is a
    draft of an OCU-`define` surface (#205); the canon freeze is the gatekeeper's.
  license:
    name: FSL-1.1-Apache-2.0
    url: https://fsl.software/

# No `servers` block. The operator plane is a host-owned local transport (the 0700
# operator socket), not a routable network origin: the binding is a deployment/IaC
# concern and the NFR-SEC-52 separation is enforced at the rendered manifest, not by a
# URL declared here. Declaring an origin would wrongly imply a network-reachable operator
# endpoint. (x-ocu-design: the absence is deliberate, not an omission.)

# Host-attested transport authentication. The operator transport (the host-owned 0700
...
# (NFR-SEC-43).
security:
  - operatorPeerCred: []
paths:
  /v1alpha/sessions:
    post:
      operationId: createSession
      summary: Create a session (operator scope).
      description: >-
        Run the lifecycle create pipeline for an operator-channel caller. The caller is
        host-attested from the operator transport's peer credential; the body fields are
        HINTS carried onto the create input (a session_hint, an image ref, mount/egress/
        resource intents, the staged control public key), never identity (NFR-SEC-43).
        Admission and quota run fail-closed before any host state exists; a refusal
        returns a BoundedReason and writes no reservation. The minted weak Storage-JWT is
        NEVER returned here — Control mints and pushes it on the separate mount-config
        plane; a create body carrying it would be a custody leak, so it is absent from
        both the request and the response.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateRequest'
      responses:
        '201':
          description: >-
            Session reserved. Returns the host-assigned session handle. The body carries
            no credential and no authority token — only the host-minted handle for
            correlation.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionHandle'
        '400':
          # A client-side invalid argument — a malformed request body. Deny -> 400 for an
          # invalid argument (canon #330 aligns every deny class to its status; an invalid
          # argument is 400, never 422). This replaces a stale '422': the shipped create
          # route returns 400 on a bad body (writeDecodeError) and never 422.
          $ref: '#/components/responses/Denied'
        '403':
          $ref: '#/components/responses/Denied'
        '409':
          $ref: '#/components/responses/Denied'
        '503':
          $ref: '#/components/responses/Denied'
    get:
      operationId: listSessions
      summary: List sessions (operator read scope) — enriched, read-only.
      description: >-
        Project the reservation registry as an enriched, read-only list for the operator
        console (the EnrichedLister read seam). The caller is host-attested from the
        operator transport's peer credential; there is no body and no mutating effect. By
        default only live rows (reserved, active) are returned; include_released=true adds
        released tombstones. This is a registry projection — it mints nothing, mutates
        nothing, and adds no control-plane state. Operator-sees-all: the read is
        deployment-wide, not scoped per operator (ADR-0022 accepted limitation).
      parameters:
        - $ref: '#/components/parameters/IncludeReleasedQuery'
      responses:
        '200':
          description: >-
            The enriched session list. Read-only projection of the registry; carries no
            credential and no authority token.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SessionView'
        '401':
          $ref: '#/components/responses/Denied'
        '503':
          $ref: '#/components/responses/Denied'
  /v1alpha/sessions/{key}:
    get:
      operationId: getSession
      summary: Read one session by key (operator read scope) — read-only.
      description: >-
        Project a single session row by its host-derived session key. The caller is
        host-attested from the operator transport's peer credential. A key that addresses
        no row returns 404 — a plain not-found, since this read-surface is deployment-wide
        and carries no cross-tenant secret to protect. Read-only: no body, no mutation.
      parameters:
        - $ref: '#/components/parameters/SessionKeyPath'
      responses:
        '200':
          description: The enriched session row.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionView'
        '401':
          $ref: '#/components/responses/Denied'
        '404':
          description: No session row addresses this key. No body.
        '503':
          $ref: '#/components/responses/Denied'
  /v1alpha/deployment:
    get:
      operationId: getDeployment
      summary: Read deployment-wide singletons (operator read scope) — read-only.
      description: >-
        Return the deployment-wide runtime singletons the console renders once, not
        per-session: the runtime tier and the runtime provider. Tier is a deployment-wide
        value set at deploy time (--runtime-tier), NEVER a per-row field. Read-only: no
        body, no mutation.
      responses:
        '200':
          description: The deployment-wide runtime singletons.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeploymentView'
        '401':
          $ref: '#/components/responses/Denied'
        '503':
          $ref: '#/components/responses/Denied'
  /v1alpha/sessions/destroy:
    post:
      operationId: destroySession
      summary: Destroy a session (operator scope).
      description: >-
        Run the lifecycle destroy path for an operator-channel caller. The host-attested
        caller gates the row lookup; the body `session_hint` is a HINT used only to
        address the operator's own handle (NFR-SEC-43). A foreign or unknown hint returns
        the same not-found-shaped DENY (a not-owned row is indistinguishable from a
        missing row), never another principal's teardown. Teardown is host-driven and
        idempotent — a repeat destroy of an already-gone session is a satisfied no-op.
        It is a POST with a hint body (not a DELETE on a hint path) because the hint is
        request data the host re-derives the key from, never an addressable resource id.
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DestroyRequest'
      responses:
        '200':
          description: Session destroyed (or already gone — idempotent).
        '403':
          $ref: '#/components/responses/Denied'
        '404':
          $ref: '#/components/responses/Denied'
        '503':
          $ref: '#/components/responses/Denied'
  /v1alpha/revoke/one:
    post:
      operationId: revokeOne
      summary: Revoke one session — denylist it and force-kill its live row (operator scope).
      description: >-
        Denylist a single session key and force-kill its live row. Force-kill exists ONLY
        on this operator ingress; the gateway audience cannot reach it (NFR-SEC-26,
        NFR-SEC-52). The recorded audit actor is the host-attested operator who acted —
        WHO ACTED, distinct from the revoke TARGET. The action is audit-gated: it is
        denied if its OCSF event fails to write (NFR-SEC-45). The `key` is the
        host-derived session key the operator names; `reason` is operator-supplied text.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RevokeOneRequest'
      responses:
        '200':
          description: Session revoked and its live row force-killed. Audit event recorded.
        '403':
          $ref: '#/components/responses/Denied'
        '503':
          $ref: '#/components/responses/Denied'
  /v1alpha/revoke/all:
    post:
      operationId: revokeAll
      summary: Engage deployment-wide DENY-ALL and force-kill every live row (operator scope).
      description: >-
        Engage the deployment-wide kill-switch (DENY-ALL) and force-kill every live row.
        This is the host-initiated stop, not a cooperative guest action — an unreachable
        guest control channel grants the guest no new authority and does not defeat it
        (NFR-SEC-01). It exists ONLY on this operator ingress (NFR-SEC-52). The audit
        actor is the host-attested operator; the action is denied if its OCSF event fails
        to write (NFR-SEC-45). After DENY-ALL is durable, every subsequent create is
        refused fail-closed until a lift.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RevokeAllRequest'
      responses:
        '200':
          description: DENY-ALL engaged and every live row force-killed. Audit event recorded.
        '403':
          $ref: '#/components/responses/Denied'
        '503':
          $ref: '#/components/responses/Denied'
  /v1alpha/resume/all:
    post:
      operationId: resumeAll
      summary: Lift the deployment-wide DENY-ALL (operator scope).
      description: >-
        Lift the deployment-wide kill-switch (the in-band resume of DENY-ALL), so creates
        are admitted again. This exists ONLY on this operator ingress; the gateway audience
        cannot reach it (NFR-SEC-26). The audit actor is the host-attested operator; the
        action is denied if its OCSF event fails to write (NFR-SEC-45). `reason` is
        operator-supplied text. It is the deliberate inverse of revoke/all; a parameterless
        POST with no body is accepted (the reason defaults to empty).
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResumeAllRequest'
      responses:
        '200':
          description: DENY-ALL lifted. Audit event recorded.
        '403':
          $ref: '#/components/responses/Denied'
        '503':
          $ref: '#/components/responses/Denied'
...
  /v1alpha/mcp-keys/revoke:
    post:
      operationId: revokeMCPKey
      summary: Revoke an sk-ocu- MCP API key by key_id (operator scope).
      description: >-
        Flip an MCP API key's status to revoked by its public `key_id` (ADR-0027). Revoke
        exists ONLY on this operator ingress (NFR-SEC-52). The audit actor is the
        host-attested operator; `reason` is operator-supplied audit text. The action is
        audit-gated fail-closed (NFR-SEC-45). Revoke is IDEMPOTENT: revoking an
        already-revoked or absent key_id is a satisfied no-op (200), so the surface is not a
        cross-tenant existence oracle — an unknown id is 200, never 404. After revoke,
        Control re-renders the hashed-key set and the gateway refreshes within NFR-SEC-04
        (<= 5 min), after which the key is refused with 401.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MCPKeyRevokeRequest'
      responses:
        '200':
          description: >-
            Revoke applied (or a satisfied no-op for an already-revoked/absent id). Audit
            event recorded. No body.
        '400':
          # A client-side invalid argument — a malformed body that did not decode. Deny
          # -> 400 for an invalid argument (canon #330 aligns every deny class to its
          # status; an invalid argument is 400, never 422).
          $ref: '#/components/responses/Denied'
        '401':
          $ref: '#/components/responses/Denied'
        '403':
          $ref: '#/components/responses/Denied'
        '503':
          $ref: '#/components/responses/Denied'

# NOTE: the operator denylist-edit (lift one deny) and quota-override are COMPLETE,
# tested in-process operator handlers (operator.Handlers.LiftDeny / .OverrideQuota)
# whose HTTP route is deliberately NOT mounted yet — the same design-fenced class as
# the SOAR-revoke pair. They are intentionally ABSENT from this spec because this
# document describes only the routes the operator socket actually serves
# (TestDeferredHandlers_AllowListIsExact enforces that they stay unmounted until their
# wire route lands deliberately). When their routes are mounted, their path items are
# added here in the same change.
components:
  securitySchemes:
    operatorPeerCred:
      type: mutualTLS
      description: >-
        Host-attested operator identity. The operator transport is the host-owned 0700
        operator socket; the caller is derived from its peer credential (SO_PEERCRED), and
        possession of an operator-scoped connection IS the operator credential (ADR-0004
        fixes the substrate). The `mutualTLS` type is the closest OpenAPI security-scheme
        token for "the transport peer's verified identity is the authority"; the v1
        substrate is a local UDS peer-cred, not network mTLS. No body field and no bearer
        token carries identity on this surface — a request body is hints only (NFR-SEC-43).
...